All the level 4 warnings were resolved in the library.

+++ AbstractNetworkCore/Sigmoids.cpp
@@ -10,10 +10,10 @@ Type mnn::root_sigmoid(Type x) {
 	return x / sqrtf(1 + x * x);
 }
 Type mnn::atan_sigmoid(Type x) {
-	return 2.f / M_PI * atanf(x * M_PI / 2.f);
+	return 2.f / float(M_PI) * atanf(x * float(M_PI) / 2.f);
 }
 Type mnn::erf_sigmoid_derivative(Type x) {
-	return 2.f * expf(-x * x) / sqrtf(M_PI);
+	return 2.f * expf(-x * x) / sqrtf(float(M_PI));
 }
 Type mnn::tanh_sigmoid_derivative(Type x) {
 	Type t = tanhf(x);
@@ -26,5 +26,5 @@ Type mnn::root_sigmoid_derivative(Type x) {
 	return -x * x / powf(1 + x * x, 3.f / 2.f) + 1.f / sqrtf(1 + x * x);
 }
 Type mnn::atan_sigmoid_derivative(Type x) {
-	return 1.f / (1 + M_PI * M_PI * x * x / 4.f);
+	return 1.f / (1 + float(M_PI * M_PI) * x * x / 4.f);
 }
